Update Light Track Testing
Smoothed out bugs and stops really well when it gets close to the light. Only thing to change is make it smoother in approaching and try to fix the fact that it biases to the left. Tried to do that by alternating the direction of the spins. I think it'll work but needs fine tuning. Wheel kept coming off of the gear so called it a night before lab tomorrow.